moving share components into component file
