+++ src/app/shared/components/confirm-dialog/confirm-dialog.component.scss
