+++ src/app/shared/components/error/error.component.spec.ts
