+++ src/app/shared/components/error/error.component.ts
