+++ src/app/shared/components/loading/loading.component.ts
