+++ src/app/shared/components/snackbar/snackbar.component.scss
