+++ src/app/shared/components/snackbar/snackbar.component.ts
@@ -1,6 +1,5 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatSnackBar, MatSnackBarRef, MAT_SNACK_BAR_DATA } from '@angular/material/snack-bar';
-import { MessageService } from '@core/services/message.service';
 
 @Component({
   selector: 'app-snackbar',
