+++ src/environments/environment.ts
@@ -4,8 +4,8 @@
 
 export const environment = {
   production: false,
-  apiUrl: 'https://squacapi.pnsn.org/',
-  // apiUrl: 'https://staging-squacapi.pnsn.org/',
+  // apiUrl: 'https://squacapi.apnsn.org/',
+  apiUrl: 'https://staging-squacapi.pnsn.org/',
   version: 'v1.0/'
 };
